update 1.0.1
update 1.0.1

+++ m2srccf/Metin2Client/UserInterface/GameType.h
@@ -339,7 +339,7 @@ enum
 typedef struct SQuickSlot
 {
 	BYTE Type;
-	BYTE Position;
+	uint16_t Position;
 } TQuickSlot;
 
 typedef struct TPlayerItemAttribute
